security groups: fix ACL creation at ovn controller
The ACLs must reference the port groups by name. OVN currently
rejects the ACL if the group name is in UUID format. This only
manifests when the OVS rules are being evaluated, with the
following warning shown in the ovn-controller log file:

"""
parsing match "!ct.new && ct.est && !ct.rpl && \
  ct_label.blocked == 0 && \
  (inport == @7f6948ce-593f-4150-803f-0a980c8147b0 && ip4)": `@' \
  must be followed by a valid identifier.
"""

This patch addresses this, by creating the security groups internal
name as strings in the "ovirt_{gen_id}" format.

Change-Id: If8f186e56b2d06d2e86e9a9407088d4d0ced3e89
Bug-Url: https://bugzilla.redhat.com/1744235
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>