integration tests: remote group id setting with new sec group
Add a test where 2 security groups are created: the icm_group,
and the limited_access group.

The limited access group is added a rule, where only members of
the icmp_group can communicate with members of the limited_access
group.

The client namespace is initially set as a member of the
limited_access group, and as such, it cannot ping the server
namespace.

Afterwards, the client namespace is set as a member of the
icmp_group, and as a result, it can ping the limited_access group.

Change-Id: I8c01f6aa5fedc74fb0e5c2f725665f75412dcfd9
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>