security groups: map the remote group ID when needed
As per the mapping changes originated from the security group
transaction refactor, and the fact that OVN does not accept port
group names in UUID format, another mapping mechanism had to be
used.

This commit addresses the missing part of the mapping, which is
the remote group ID parameter in the security group rules.

In order to make it work, it is needed to read the remote group,
addressing it by name, extract the remote group ID, and wrap the
ACL object with the remote group ID referred in the rule.

Change-Id: I2b9022ce91c6d220a42a9398f4661b1c1da67a41
Bug-Url: https://bugzilla.redhat.com/1744235
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>