docs: explain the security group ID mapping update
The security group ID in both security groups and security group
rules must match, both in the OVN 'world' and in the
'networking api' world for the security group use cases to be
fulfilled.

This documents explains the plan to achieve that goal, while
preserving backwards compatibility with security groups previously
created.

Change-Id: Iac9a413e417a6c5e3dd003f59bb59a79516e4ab2
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>