integration tests: correct the used security groups
The scenario is currently setting the 'icmp_client' to be a member
of the Default security group, while the 'icmp_server' is not a
member of any group.

The Default group has rules that - on packet ingress - check if the
packet 'originator' is a member of the default group. Since the
ping is sent from client to server, the Default group membership
on egress is not checked, and as a result, the ping succeeds.

Inverting the group membership - i.e. have the 'icmp_server' as
member of the Default group corrects the test case.

Change-Id: I4b19d462c26ebfb9563bcb37a1e897359d3fb096
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>