integration-tests: always create the artifacts dir
Up to now, the artifacts dir was meant to have only container
information.

As such, there was no point in creating it *if* the containers
were not successfuly run.

We now plan to export also host runtime info - e.g. podman info -
to improve debuggability of the integration test environment.

Change-Id: Ia6eaeade1d4ccb5b906d18a0a26b8ea9543abcfc
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>