Improve scaling of port listing
Most expensive step during listing the ports is to find the
networks which is associated with the port.
This can be avoided, by iterating over all ports of all
networks.
While this changes the order of the ports in the list of
ports, it still covers all ports, because every port is
associated with a network.

A scenario listing the ports of 750 networks with two ports
each, took around 19 seconds before, takes around 2 seconds
with this fix.

A scenario listing the ports of a single network with 1500
ports, took around 36 seconds before, takes around 2 seconds
with this fix.

Change-Id: I65506eb2709413c5e42dd3ead59c277e6e74a507
Bug-Url: https://bugzilla.redhat.com/1835550
Signed-off-by: Dominik Holler <dholler@redhat.com>