Improve consistency in unit test scenario
Before the change, the mocked port list misses a port which is
referenced by a network, which is referenced by the mocked
list of networks. This way the set of all ports of all
networks did not match the set of all ports.

The mocked port list should contain all three ports which are
used in the mocked networks, to have an consistent state of
the mocked db.
The consistent state of the db will be required by the
following change.

Change-Id: I142744f88fff7e5ec2de786f50b965ccc2a5434b
Signed-off-by: Dominik Holler <dholler@redhat.com>