Get reference to dhcp directly from lsp
If a lsp has a reference to dhcp options, this reference is
used to report the subnet. If a lsp, e.g. a router port, has
not such a reference, the subnet is derived from the ls
associated with the lsp. The latter slows down with the number
of subnets in the db, but impacts only router ports.

This change improves scaling of listing ports associated to a
subnet:
A scenario of listing 1k ports spread over 500 networks each
with a subnet which would take 36 seconds is reduced by this
change to less than 1 second.

Change-Id: I31a38a119f744e6fa93030f8f9db902ccf40840b
Bug-Url: https://bugzilla.redhat.com/1835550
Signed-off-by: Dominik Holler <dholler@redhat.com>